testing strategies for ejs
